// lib/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  student: 'student',
  council: 'council',
  admin: 'admin'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const UserStatus = {
  active: 'active',
  pending: 'pending',
  suspended: 'suspended'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const BoardType = {
  official: 'official',
  community: 'community'
} as const

export type BoardType = (typeof BoardType)[keyof typeof BoardType]


export const PostType = {
  official: 'official',
  community: 'community'
} as const

export type PostType = (typeof PostType)[keyof typeof PostType]


export const TargetType = {
  post: 'post',
  comment: 'comment'
} as const

export type TargetType = (typeof TargetType)[keyof typeof TargetType]


export const ReportStatus = {
  pending: 'pending',
  reviewed: 'reviewed',
  dismissed: 'dismissed'
} as const

export type ReportStatus = (typeof ReportStatus)[keyof typeof ReportStatus]


export const MealType = {
  breakfast: 'breakfast',
  lunch: 'lunch',
  dinner: 'dinner'
} as const

export type MealType = (typeof MealType)[keyof typeof MealType]